requirements.txt
gym-super-mario-bros==7.4.0
nes-py==8.2.1
stable-baselines3==2.1.0
torch>=2.0.0
gym==0.26.2
tensorboard>=2.14.0
opencv-python>=4.8.0
numpy>=1.24.0
rich>=13.7.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pyyaml>=6.0.1
pillow>=10.2.0
pydantic>=2.6.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
